@@ -1,7 +1,7 @@
 {
   "name": "next-redux-wrapper-demo",
   "private": true,
-  "version": "3.0.0-alpha.3",
+  "version": "3.0.0-alpha.4",
   "description": "Demo of redux wrapper for Next.js",
   "scripts": {
     "clean": "rimraf .next coverage",
@@ -11,7 +11,7 @@
     "serve": "next start"
   },
   "dependencies": {
-    "next-redux-wrapper": "^3.0.0-alpha.3",
+    "next-redux-wrapper": "^3.0.0-alpha.4",
     "react": "16.8.6",
     "react-dom": "16.8.6",
     "react-redux": "7.0.3",
@@ -21,16 +21,15 @@
     "@types/expect-puppeteer": "3.3.1",
     "@types/jest": "24.0.13",
     "@types/jest-environment-puppeteer": "4.0.0",
-    "@types/next": "8.0.5",
     "@types/puppeteer": "1.12.4",
     "@types/react": "16.8.19",
+    "@types/react-dom": "16.9.0",
     "@types/react-redux": "7.0.9",
     "@types/redux-promise-middleware": "0.0.11",
     "@types/webpack-env": "1.13.9",
-    "@zeit/next-typescript": "1.1.1",
     "jest-puppeteer": "4.2.0",
-    "next": "8.1.0",
-    "next-redux-wrapper-configs": "^3.0.0-alpha.3",
+    "next": "9.0.3",
+    "next-redux-wrapper-configs": "^3.0.0-alpha.4",
     "puppeteer": "1.17.0",
     "rimraf": "2.6.3",
     "ts-jest": "24.0.2",
