@@ -1,6 +1,6 @@
 {
   "name": "next-redux-wrapper",
-  "version": "3.0.0-alpha.3",
+  "version": "3.0.0-alpha.4",
   "description": "Redux wrapper for Next.js",
   "main": "lib/index.js",
   "module": "es6/index.js",
@@ -19,13 +19,12 @@
   },
   "devDependencies": {
     "@types/jest": "24.0.13",
-    "@types/next": "8.0.5",
     "@types/react": "16.8.19",
     "@types/react-redux": "7.0.9",
     "@types/redux-promise-middleware": "0.0.11",
-    "@zeit/next-typescript": "1.1.1",
     "jest": "24.8.0",
-    "next-redux-wrapper-configs": "^3.0.0-alpha.3",
+    "next": "9.0.3",
+    "next-redux-wrapper-configs": "^3.0.0-alpha.4",
     "npm-run-all": "4.1.5",
     "react": "16.8.6",
     "react-dom": "16.8.6",
@@ -38,7 +37,7 @@
     "typescript": "3.5.1"
   },
   "peerDependencies": {
-    "next": ">=6.0.0",
+    "next": ">=9.0.0",
     "react": "*"
   },
   "author": "Kirill Konshin",
