@@ -1,4 +1,4 @@
-const config = require('../jest-puppeteer.config');
+import config from '../jest-puppeteer.config';
 
 const openPage = (url = '/') => page.goto(`http://localhost:${config.server.port}${url}`);
 
