@@ -1,2 +1,1 @@
-const withTypescript = require('@zeit/next-typescript');
-module.exports = withTypescript();
+module.exports = {};
